+++ pepper/src/buffer.rs
@@ -2306,4 +2306,3 @@ mod tests {
         assert_eq!(3, len(&buffer, 2));
     }
 }
-