+++ pepper/src/events.rs
@@ -1009,4 +1009,3 @@ mod tests {
         assert!(parser.next().is_none());
     }
 }
-